+++ ReleaseNotes/release_notes.md
@@ -4,6 +4,7 @@
   1. improve `parallel_copy`,  rely on `xargs` when `GNU parallel` is not available
   1. add add parallel `sync` based on `rsync -a` & `xargs/parallel`
   1. `sqlitedict` : tiny fix for warning in `python-3.11`
+  1. fix `pypdf` examples
   
 ## Backward incompatible